# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "These are the APIs for VideoVerse take home assignment.",
        "title": "VideoVerse API",
        "contact": {
            "name": "Khush Patibandha",
            "email": "khush.patibandha@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/uploads/{link}": {
            "get": {
                "description": "Given a valid temporary link, redirects to the video file",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "video"
                ],
                "summary": "Redirect to the uploaded video",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Temporary video link",
                        "name": "link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video file served"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}